requirements.txt
Flask
flask-cors
Pillow
cryptography
gunicorn